+++ src/world/objects/Point3D.java
@@ -4,11 +4,11 @@ import java.security.PublicKey;
 
 public class Point3D {
 
-    int x;
-    int y;
-    int z;
+    double x;
+    double y;
+    double z;
 
-    public Point3D(int x, int y, int z){
+    public Point3D(double x, double y, double z){
         this.x = x;
         this.y = y;
         this.z = z;
@@ -24,20 +24,20 @@ public class Point3D {
     }
 
 
-    public int getX(){
+    public double getX(){
         return x;
     }
 
-    public int getY(){
+    public double getY(){
         return y;
     }
 
-    public int getZ(){
+    public double getZ(){
         return z;
     }
 
-    public void setX(int x){this.x = x;}
-    public void setY(int y){this.y = y;}
-    public void setZ(int z){this.z = z;}
+    public void setX(double x){this.x = x;}
+    public void setY(double y){this.y = y;}
+    public void setZ(double z){this.z = z;}
 
 }
